requirements.txt
Flask
openai
python-dotenv
opencv-python
numpy
pillow
torch
torchvision
requests
bs4
